+++ src/main/kotlin/io/github/cotrin8672/cel/content/SharedStorageBehaviour.kt
@@ -13,8 +13,11 @@ import io.github.cotrin8672.cel.registry.CelItems
 import io.github.cotrin8672.cel.util.CelLang
 import io.github.cotrin8672.cel.util.StorageFrequency
 import io.github.cotrin8672.cel.util.storageFrequency
+import io.github.cotrin8672.cel.util.use
 import net.createmod.catnip.math.VecHelper
 import net.minecraft.ChatFormatting
+import net.minecraft.client.gui.Font
+import net.minecraft.client.gui.GuiGraphics
 import net.minecraft.core.Direction
 import net.minecraft.nbt.CompoundTag
 import net.minecraft.network.chat.Component
@@ -27,6 +30,7 @@ import net.minecraft.world.entity.player.Player
 import net.minecraft.world.item.ItemStack
 import net.minecraft.world.phys.BlockHitResult
 import net.minecraft.world.phys.Vec3
+import net.minecraftforge.client.IItemDecorator
 import kotlin.math.max
 
 open class SharedStorageBehaviour(
@@ -35,6 +39,38 @@ open class SharedStorageBehaviour(
 ) : BlockEntityBehaviour(be), ValueSettingsBehaviour {
     companion object {
         val TYPE = BehaviourType<SharedStorageBehaviour>()
+
+        val DECORATOR = IItemDecorator { guiGraphics: GuiGraphics, font: Font, stack: ItemStack, x: Int, y: Int ->
+            val storageFrequency = stack.storageFrequency
+            val frequencyItem = storageFrequency.stack
+            if (frequencyItem.isEmpty) false
+            guiGraphics.pose().use {
+                val xOffset = x + 15f
+                val yOffset = y + 1f
+                translate(xOffset, yOffset, 0f)
+                scale(0.5f, 0.5f, 1f)
+                translate(-xOffset, -yOffset, 100f)
+                guiGraphics.renderItem(
+                    if (storageFrequency.isGlobalScope)
+                        storageFrequency.stack
+                    else
+                        CelItems.SCOPE_FILTER.asStack(), x, y
+                )
+
+                if (storageFrequency.isPersonalScope) {
+                    use {
+                        val xOffset = x + 8f
+                        val yOffset = y + 8f
+                        translate(xOffset, yOffset, 0f)
+                        scale(0.5f, 0.5f, 1f)
+                        translate(-xOffset, -yOffset, 10f)
+                        guiGraphics.renderItem(storageFrequency.stack, x, y)
+                    }
+                }
+            }
+
+            true
+        }
     }
 
     private var storageFrequency: StorageFrequency = StorageFrequency.EMPTY
@@ -98,6 +134,12 @@ open class SharedStorageBehaviour(
 
     open fun getRenderDistance(): Float {
         return AllConfigs.client().filterItemRenderDistance.f
+    }
+
+    fun setStorageFrequency(storageFrequency: StorageFrequency) {
+        this.storageFrequency = storageFrequency
+        blockEntity.setChanged()
+        blockEntity.sendData()
     }
 
     open fun setFrequencyItem(face: Direction?, stack: ItemStack): Boolean {