+++ src/main/kotlin/io/github/cotrin8672/cel/datagen/CelRecipeProvider.kt
@@ -6,11 +6,7 @@ import io.github.cotrin8672.cel.CreateEnderLink
 import io.github.cotrin8672.cel.registry.CelBlocks
 import io.github.cotrin8672.cel.registry.CelItems
 import net.minecraft.data.PackOutput
-import net.minecraft.data.recipes.FinishedRecipe
-import net.minecraft.data.recipes.RecipeCategory
-import net.minecraft.data.recipes.RecipeProvider
-import net.minecraft.data.recipes.ShapedRecipeBuilder
-import net.minecraft.data.recipes.ShapelessRecipeBuilder
+import net.minecraft.data.recipes.*
 import net.minecraft.tags.ItemTags
 import net.minecraft.world.item.Items
 import java.util.function.Consumer
@@ -19,7 +15,7 @@ class CelRecipeProvider(
     output: PackOutput,
 ) : RecipeProvider(output) {
     override fun buildRecipes(recipeOutput: Consumer<FinishedRecipe>) {
-        with(ShapedRecipeBuilder.shaped(RecipeCategory.MISC, CelBlocks.ENDER_VAULT.asItem())) {
+        with(ShapedRecipeBuilder.shaped(RecipeCategory.MISC, CelBlocks.ENDER_VAULT)) {
             pattern(" B ")
             pattern("EVE")
             pattern(" B ")
@@ -30,7 +26,13 @@ class CelRecipeProvider(
             save(recipeOutput)
         }
 
-        with(ShapedRecipeBuilder.shaped(RecipeCategory.MISC, CelBlocks.ENDER_TANK.asItem())) {
+        with(ShapelessRecipeBuilder.shapeless(RecipeCategory.MISC, CelBlocks.ENDER_VAULT)) {
+            requires(CelBlocks.ENDER_VAULT)
+            unlockedBy("has_ender_vault", has(CelBlocks.ENDER_VAULT))
+            save(recipeOutput, CreateEnderLink.asResource("ender_vault_clear"))
+        }
+
+        with(ShapedRecipeBuilder.shaped(RecipeCategory.MISC, CelBlocks.ENDER_TANK)) {
             pattern(" B ")
             pattern("ETE")
             pattern(" B ")
@@ -41,7 +43,13 @@ class CelRecipeProvider(
             save(recipeOutput)
         }
 
-        with(ShapedRecipeBuilder.shaped(RecipeCategory.MISC, CelItems.SCOPE_FILTER.asItem())) {
+        with(ShapelessRecipeBuilder.shapeless(RecipeCategory.MISC, CelBlocks.ENDER_TANK)) {
+            requires(CelBlocks.ENDER_TANK)
+            unlockedBy("has_ender_tank", has(CelBlocks.ENDER_TANK))
+            save(recipeOutput, CreateEnderLink.asResource("ender_tank_clear"))
+        }
+
+        with(ShapedRecipeBuilder.shaped(RecipeCategory.MISC, CelItems.SCOPE_FILTER)) {
             pattern("AWA")
             define('A', Items.AMETHYST_SHARD)
             define('W', ItemTags.WOOL)
@@ -49,7 +57,7 @@ class CelRecipeProvider(
             save(recipeOutput)
         }
 
-        with(ShapelessRecipeBuilder.shapeless(RecipeCategory.MISC, CelItems.SCOPE_FILTER.asItem())) {
+        with(ShapelessRecipeBuilder.shapeless(RecipeCategory.MISC, CelItems.SCOPE_FILTER)) {
             requires(CelItems.SCOPE_FILTER)
             unlockedBy("has_scope_filter", has(CelItems.SCOPE_FILTER))
             save(recipeOutput, CreateEnderLink.asResource("scope_filter_clear"))