+++ src/main/kotlin/io/github/cotrin8672/cel/registry/CelBlocks.kt
@@ -8,6 +8,7 @@ import com.simibubi.create.foundation.data.SharedProperties
 import com.simibubi.create.foundation.data.TagGen.pickaxeOnly
 import com.tterrag.registrate.util.entry.BlockEntry
 import io.github.cotrin8672.cel.CreateEnderLink.REGISTRATE
+import io.github.cotrin8672.cel.content.block.SharedStorageBlockItem
 import io.github.cotrin8672.cel.content.block.tank.EnderTankBlock
 import io.github.cotrin8672.cel.content.block.vault.EnderVaultBlock
 import net.minecraft.client.renderer.RenderType
@@ -17,7 +18,7 @@ import java.util.function.Supplier
 
 object CelBlocks {
     val ENDER_VAULT: BlockEntry<EnderVaultBlock> = REGISTRATE
-        .block<EnderVaultBlock>("ender_vault", ::EnderVaultBlock)
+        .block("ender_vault", ::EnderVaultBlock)
         .initialProperties(SharedProperties::softMetal)
         .properties {
             it
@@ -28,22 +29,21 @@ object CelBlocks {
         .transform(pickaxeOnly())
         .blockstate(BlockStateGen.horizontalAxisBlockProvider(false))
         .transform(mountedItemStorage(CelMountedStorageTypes.SHARED_ITEM))
-        .item()
+        .item(::SharedStorageBlockItem)
         .build()
         .register()
 
     val ENDER_TANK: BlockEntry<EnderTankBlock> = REGISTRATE
-        .block<EnderTankBlock>("ender_tank", ::EnderTankBlock)
+        .block("ender_tank", ::EnderTankBlock)
         .initialProperties(SharedProperties::copperMetal)
         .properties {
-            it.noOcclusion()
-                .isRedstoneConductor { _, _, _ -> true }
+            it.noOcclusion().isRedstoneConductor { _, _, _ -> true }
         }
         .transform(pickaxeOnly())
         .blockstate { c, p -> p.simpleBlock(c.get(), AssetLookup.standardModel(c, p)) }
         .transform(mountedFluidStorage(CelMountedStorageTypes.SHARED_FLUID))
         .addLayer { Supplier(RenderType::cutoutMipped) }
-        .item()
+        .item(::SharedStorageBlockItem)
         .build()
         .register()